Using float32 to action

+++ gym_games/envs/foo_env.py
@@ -10,7 +10,7 @@ class FooEnv(gym.Env):
     def __init__(self):
         self.count = 0
         self.observation_space = spaces.Box(low=-np.inf, high=np.inf, shape=(1,), dtype=np.float32)
-        self.action_space = spaces.Box(low=0, high=20, shape=(1,), dtype=np.int8)
+        self.action_space = spaces.Box(low=0, high=20, shape=(1,), dtype=np.float32)
 
     def step(self, action):
         self.count += 1
